do modifications
