+++ app/models/assignment_team.rb
@@ -1,10 +1,10 @@
-require 'TFD1770_refactor'
+require 'file_support'
 class AssignmentTeam < Team
   belongs_to :assignment, class_name: 'Assignment', foreign_key: 'parent_id'
   has_many :review_mappings, class_name: 'ReviewResponseMap', foreign_key: 'reviewee_id'
   has_many :review_response_maps, foreign_key: 'reviewee_id'
   has_many :responses, through: :review_response_maps, foreign_key: 'map_id'
-  include Instance_method
+  include File_support
   # START of contributor methods, shared with AssignmentParticipant
 
   # Whether this team includes a given participant or not
