+++ lib/file_support.rb
@@ -1,0 +1,15 @@
+module File_support
+  def files(directory)
+    files_list = Dir[directory + "/*"]
+    files = []
+
+    files_list.each do |file|
+      if File.directory?(file)
+        dir_files = files(file)
+        dir_files.each {|f| files << f }
+      end
+      files << file
+    end
+    files
+  end
+end
